Cleaned up the repository
