+++ src/classes/VirtualMachine.ts
@@ -1,0 +1,7 @@
+export class VirtualMachine {
+    id: string;
+    name: string;
+    powerState: string;
+    resourceGroupName: string;
+    hardwareProfile: { vmSize: string };
+}
