+++ src/classes/Workflow.ts
@@ -1,0 +1,12 @@
+export class Workflow {
+    nodes: string;
+    edges: string;
+    owner: string;
+    name: string;
+    id: string;
+    created: string;
+    status: string;
+    todo_nodes: string[];
+    busy_nodes: string[];
+    done_nodes: string[];
+}
