+++ src/repositories/TaskRepository.ts
@@ -1,73 +1,15 @@
 import fetch from 'node-fetch';
 
 import { Task } from '../classes/Task';
-
-class Workflow {
-    nodes: string;
-    edges: string;
-    owner: string;
-    name: string;
-    id: string;
-    created: string;
-    status: string;
-    todo_nodes: string[];
-    busy_nodes: string[];
-    done_nodes: string[];
-}
-
-
-
-function filterTaskIsHuman(task: Task): boolean {
-    const taskInfo = task.task_id.split(':');
-    return (taskInfo.length === 3 && (taskInfo[1] === 'HE' || taskInfo[1] === 'HH'));
-}
-
-function filterTaskIsWorker(task: Task): boolean {
-    return !filterTaskIsHuman(task);
-}
-
-function filterTaskIsTodo(task: Task): boolean {
-    return task.task_status === 'todo';
-}
-
-function filterTaskIsBusy(task: Task): boolean {
-    return task.task_status === 'busy';
-}
-
-function filterTaskIsDone(task: Task): boolean {
-    return task.task_status === 'done';
-}
-
-function getPreviousTaskIds(task: Task, edges_string: string): string[] {
-    const previous_tasks: string[] = [];
-
-    const edge_words = edges_string.split(',').map(w => w.trim());
-    edge_words.forEach((word) => {
-        const elements = word.split('->').map(w => w.trim());
-        if (task.task_id === elements[1]) {
-            previous_tasks.push(elements[0]);
-        }
-    });
-
-    return previous_tasks;
-};
-
-function filterTaskIsFree(task: Task, workflow: Workflow): boolean {
-    if (task.task_status !== 'todo') {
-        return false;
-    }
-
-    const previous_tasks: string[] = getPreviousTaskIds(task, workflow.edges);
-    let taskIsFree = true;
-
-    previous_tasks.forEach((t) => {
-        if (workflow.done_nodes.indexOf(t) === -1) {
-            taskIsFree = false;
-        }
-    });
-
-    return taskIsFree;
-}
+import { Workflow } from '../classes/Workflow';
+import {
+    filterTaskIsHuman,
+    filterTaskIsBusy,
+    filterTaskIsDone,
+    filterTaskIsFree,
+    filterTaskIsTodo,
+    filterTaskIsWorker
+} from './filters/FilterFunctions';
 
 export class TaskRepository {
     host = process.env.API || 'localhost';
